+++ cmd/logger_unix.go
@@ -1,0 +1,5 @@
+package cmd
+
+func toZapLogPath(p string) string {
+	return p
+}