+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### 🐛 Fixes
 - [558](https://github.com/vegaprotocol/vegawallet/pull/558) - Use user-friendly RPC code as error message
+- [569](https://github.com/vegaprotocol/vegawallet/pull/569) - Ensure the winfile scheme for logger is register only once
 
 ## 0.15.0